Make ECS library a comptime-configurable struct to support multiple ECS engines (#24)
* Add configurable ECS backend support

Implement comptime-configurable ECS backend abstraction (closes #7):

- Add src/ecs/interface.zig: Central ECS interface that exports
  Registry and Entity types based on build option selection

- Add src/ecs/zig_ecs_adapter.zig: Adapter wrapping prime31/zig-ecs
  to conform to the interface contract

- Add src/ecs/zcs_adapter.zig: Stub adapter for ZCS backend
  (ZCS v0.1.1 has @Tuple compatibility issue with Zig 0.15+)

- Update build.zig: Add -Decs_backend option (zig_ecs|zcs)
  to select ECS implementation at compile time

- Update build.zig.zon: Document ZCS dependency (commented out
  until Zig version compatibility is resolved)

The abstraction is transparent to existing code - files continue
to use @import("ecs") which now resolves to the interface module.
All 105 tests pass with both backends.



* Add example_4 demonstrating ECS backend configuration

- Add usage/example_4/ with bouncing shapes demo
- Fix ECS interface to use correct build_options type names
- Fix ZCS adapter to use comptime type name hashing

The example shows how to:
- Use `-Decs_backend=zig_ecs` or `-Decs_backend=zcs` build options
- Write backend-agnostic ECS code using the interface
- Registry operations (create, add, tryGet) work identically on all backends



* Add example_4 to CI and update generator template with ECS backend

- Update CI to generate and run example_4 with both ECS backends
- Update build_zig.txt template to include ECS backend option



* Address PR review comments for ZCS adapter

- Fix usage example to use @import("ecs") instead of path
- Add clear documentation about memory leak behavior in stub
- Fix duplicate component add to free previous allocation
- Improve comments explaining stub limitations



* Replace ZCS with zflecs backend and add ECS benchmarks

- Remove zcs_adapter (not compatible with Zig 0.15)
- Add zflecs_adapter with full 64-bit entity support
- Add comprehensive ECS benchmark suite
- Make Entity type generic (32-bit for zig_ecs, 64-bit for zflecs)
- Add optional Entity.invalid and Entity.eql with NotImplemented fallback
- Export entity_size and capability flags in interface

Benchmark highlights:
- zig_ecs: ~500M ops/s for component lookup, ~110M ops/s for view iteration
- zflecs: ~11M ops/s for component lookup, ~10B ops/s for view iteration
- zig_ecs better for random access, zflecs better for bulk iteration



* Add ecs_backend option to project.labelle configuration

Users can now select the ECS backend in their project.labelle file:

  .{
      .name = "my_game",
      .ecs_backend = .zflecs,  // or .zig_ecs (default)
      ...
  }

- Add EcsBackend enum (zig_ecs, zflecs) to project_config.zig
- Update build_zig.txt template to use config's ecs_backend as default
- Fix template to use zflecs instead of removed zcs backend
- Generator now passes ecs_backend to template



* Support 64-bit entities for zflecs backend

- Update lifecycle hooks to use u64 instead of u32 for entity parameters
- Add entityToU64/entityFromU64 helper functions in scene.zig
- Fix render_pipeline.zig to truncate entity bits for graphics layer
- Update prefab.zig to use u64 in Prefab struct
- Replace remaining zcs references with zflecs in CI and build.zig
- Update example_4 files to reference zflecs backend



---------

+++ src/prefab.zig
@@ -32,14 +32,15 @@ pub const SpriteConfig = struct {
 };
 
 /// Type-erased prefab interface for runtime use
-/// Uses u32 for entity and *anyopaque for Game to avoid circular imports
+/// Uses u64 for entity and *anyopaque for Game to avoid circular imports.
+/// The u64 type accommodates both 32-bit (zig_ecs) and 64-bit (zflecs) entity IDs.
 pub const Prefab = struct {
     name: []const u8,
     sprite: SpriteConfig,
     animation: ?[]const u8,
-    onCreate: ?*const fn (u32, *anyopaque) void,
-    onUpdate: ?*const fn (u32, *anyopaque, f32) void,
-    onDestroy: ?*const fn (u32, *anyopaque) void,
+    onCreate: ?*const fn (u64, *anyopaque) void,
+    onUpdate: ?*const fn (u64, *anyopaque, f32) void,
+    onDestroy: ?*const fn (u64, *anyopaque) void,
 };
 
 /// Check if a type is a valid prefab
